Fix serious bug in MCDF-XCOM PE routine
Unexpected behavior: Results depend on the order in which the material
composition is defined.

Reported by Iymad Mansour who observed that egs_brachy simulations of
an irradiation with an isotropic 35 keV photon point source embedded
in a steel applicator, produced 5% differences in the surface dose of
a water phantom when the order of the steel components was changed.

Confirmed with egs_kerma for a 35 keV photon point source surrounded by
a 1 mm steel encapsulation embedded in a spherical water phantom. Kerma
at the spherical water shell next to the encapsulation differs by 4%.

BUG: Sampling element and atomic shell for photo-electric interaction
     with MCDF-XCOM shell-wise using unsorted array zelem rather than
     pe_zsorted.

Using pe_zsorted array fixes the problem.